fixup int return

+++ codonW.h
@@ -178,6 +178,6 @@ int hydro_out(FILE *foutput, long *naa, MENU_STRUCT *pm);
 int aromo_out(FILE *foutput, long *naa, MENU_STRUCT *pm);
 int cutab_out(FILE *fblkout, long *nncod, long *nnaa, MENU_STRUCT *pm);
 int dinuc_out(FILE *fblkout, char *title, char sep);
-float enc_out(FILE *foutput, long *ncod, long *naa, MENU_STRUCT *pm);
+int enc_out(FILE *foutput, long *ncod, long *naa, MENU_STRUCT *pm);
 int gc_out(FILE *foutput, FILE *fblkout, int which, MENU_STRUCT *pm);
 int base_sil_us_out(FILE *foutput, long *ncod, long *naa, MENU_STRUCT *pm);
